% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalFrOdo.R
\name{PalFrOdo}
\alias{PalFrOdo}
\title{Custom palettes}
\usage{
PalFrOdo(name)
}
\arguments{
\item{name}{Name of the palette}
}
\value{
A vector with colors of the selected palette
}
\description{
Custom palettes
}
